pyproject.toml
[project]
name = "atp"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "gallery-dl>=1.29.6",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.40",
    "tiktokapi>=7.1.0",
    "yt-dlp>=2025.4.30",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
